+++ custom_components/localthings/registry/by_type/dryer.py
@@ -9,21 +9,15 @@ objects the washer registry uses -- washer and dryer expose the same
 DA_WM_-family surface, so they stay consistent instead of each carrying a
 bespoke variant.
 """
-from ..capabilities import common, dryer, fridge, ignored, laundry, operational
+from ..capabilities import common, dryer, ignored, laundry, operational
 from ._base import DeviceRegistry, _build
 
 REGISTRY = DeviceRegistry(
     name='dryer',
     capabilities=_build([
         *ignored.IGNORED,
-        common.POWER_GENERIC,
-        common.POWER_VS_FALLBACK,
-        common.KIDS_LOCK_GENERIC,
-        common.KIDS_LOCK_VS_FALLBACK,
-        common.REMOTE_CONTROL_GENERIC,
-        common.REMOTE_CONTROL_VS_FALLBACK,
-        common.ALARMS,
-        common.ENERGY_METER,
+        *common.UNIVERSAL,
+        *common.POWER,
         operational.OPERATIONAL_STATE,
         laundry.DOOR_LED,
         laundry.SOUND_MODE,
@@ -32,6 +26,5 @@ REGISTRY = DeviceRegistry(
         dryer.DRYER_SETTINGS,
         dryer.DRYER_COURSE,
         dryer.DRYER_DIAGNOSIS,
-        fridge.FIRMWARE_UPDATE,
     ]),
 )
