+++ custom_components/localthings/registry/by_type/refrigerator.py
@@ -6,21 +6,14 @@ REGISTRY = DeviceRegistry(
     name='refrigerator',
     capabilities=_build([
         *ignored.IGNORED,
-        common.POWER_GENERIC,
-        common.POWER_VS_FALLBACK,
+        *common.UNIVERSAL,
+        *common.POWER,
         fridge.STATUS_LOCK,
         fridge.DOOR_ALERT,
-        common.KIDS_LOCK_GENERIC,
-        common.KIDS_LOCK_VS_FALLBACK,
-        common.REMOTE_CONTROL_GENERIC,
-        common.REMOTE_CONTROL_VS_FALLBACK,
-        common.ALARMS,
-        common.ENERGY_METER,
         common.WATER_FILTER,
         dishwasher.DIAGNOSIS,
         fridge.ICEMAKER_NIGHTTIME,
         fridge.FLEX_ZONE,
-        common.AI_ENERGY_LEVEL,
         fridge.REFRIGERATION,
         fridge.AUTOFILL,
         fridge.WELCOME_LIGHTING,
@@ -29,11 +22,9 @@ REGISTRY = DeviceRegistry(
         fridge.SABBATH,
         fridge.BEVERAGE_ZONE,
         fridge.PANTRY_ZONE,
-        fridge.FIRMWARE_UPDATE,
         fridge.DEFROST_DELAY,
         fridge.DEFROST_DELAY_NATIVE_DUPLICATE,
         fridge.DEFROST_BLOCK_STATUS,
-        fridge.SELF_CHECK,
         fridge.DOORS_FALLBACK,
         fridge.TEMPERATURES_FALLBACK,
         fridge.ICEMAKER_STATUS_FALLBACK,
